requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
async-timeout==4.0.2
    # via
    #   -r requirements/base.txt
    #   redis
attrs==22.2.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
certifi==2022.12.7
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
cryptography==39.0.2
    # via
    #   -r requirements/base.txt
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
django==3.2.18
    # via
    #   -r requirements/base.txt
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   drf-yasg
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-rbac
    #   jsonfield2
    #   openedx-events
    #   openedx-ledger
django-cors-headers==3.14.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-rbac
django-extensions==3.2.1
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
django-filter==22.1
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
django-model-utils==4.3.1
    # via
    #   -r requirements/base.txt
    #   edx-rbac
django-simple-history==3.0.0
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangoql==0.17.1
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   drf-yasg
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
drf-yasg==1.21.5
    # via -r requirements/base.txt
edx-auth-backends==4.1.0
    # via -r requirements/base.txt
edx-django-release-util==1.2.0
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
edx-django-utils==5.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   openedx-ledger
edx-drf-extensions==8.4.1
    # via
    #   -r requirements/base.txt
    #   edx-rbac
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   openedx-events
edx-rbac==1.7.0
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
edx-rest-api-client==5.5.0
    # via -r requirements/base.txt
fastavro==1.7.3
    # via
    #   -r requirements/base.txt
    #   openedx-events
future==0.18.3
    # via
    #   -r requirements/base.txt
    #   pyjwkest
gevent==22.10.2
    # via -r requirements/production.in
greenlet==2.0.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements/production.in
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   coreschema
jsonfield2==4.0.0.post0
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
markupsafe==2.1.2
    # via
    #   -r requirements/base.txt
    #   jinja2
mysqlclient==2.1.1
    # via
    #   -r requirements/base.txt
    #   -r requirements/production.in
    #   openedx-ledger
newrelic==8.7.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
openedx-events==7.0.0
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
openedx-ledger==0.2.0
    # via -r requirements/base.txt
packaging==23.0
    # via
    #   -r requirements/base.txt
    #   drf-yasg
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
ply==3.11
    # via
    #   -r requirements/base.txt
    #   djangoql
psutil==5.9.4
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.17
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
python-memcached==1.59
    # via -r requirements/production.in
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2022.7.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
    #   drf-yasg
    #   openedx-ledger
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/production.in
    #   edx-django-release-util
redis==4.5.1
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
requests==2.28.2
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
ruamel-yaml==0.17.21
    # via
    #   -r requirements/base.txt
    #   drf-yasg
ruamel-yaml-clib==0.2.7
    # via
    #   -r requirements/base.txt
    #   ruamel-yaml
rules==3.3
    # via
    #   -r requirements/base.txt
    #   openedx-ledger
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-rbac
    #   pyjwkest
    #   python-dateutil
    #   python-memcached
slumber==0.7.1
    # via
    #   -r requirements/base.txt
    #   edx-rest-api-client
social-auth-app-django==5.1.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
social-auth-core==4.4.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-opaque-keys
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.15
    # via
    #   -r requirements/base.txt
    #   requests
zope-event==4.6
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
